Updated scala code to use the new ensemble member definition code.

+++ src/framework/io/serialize/AbstractEvaluatorSerializer.scala
@@ -6,7 +6,6 @@ import fieldml.valueType.ValueType
 import util.exception._
 
 import fieldml.jni.FieldmlApi._
-import fieldml.jni.TypeBoundsType
 import fieldml.jni.FieldmlHandleType._
 import fieldml.jni.FieldmlApiConstants._
 
@@ -17,7 +16,7 @@ import framework.valuesource.AbstractEvaluatorValueSource
 
 object AbstractEvaluatorSerializer
 {
-    def insert( handle : Long, evaluator : AbstractEvaluator ) : Unit =
+    def insert( handle : Int, evaluator : AbstractEvaluator ) : Unit =
     {
         val typeHandle = GetNamedObject( handle, evaluator.valueType.name )
         val objectHandle = Fieldml_CreateAbstractEvaluator( handle, evaluator.name, typeHandle )
